scripts/SQL/requirements.txt
psycopg2-binary == 2.9.9
configparser == 7.1.0
colorama == 0.4.6